third_party/WebKit/LayoutTests/fast/dom/Range/insertNode-trigger-onpageshow-crash.html

<!DOCTYPE html>
<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>
<audio src="aOyer">
<script>
  const audio = document.querySelector('audio');
  const newNode = document.createElement('frame');

  function handler(event) {
    document.body.webkitRequestFullscreen();
    var range = document.createRange();
    range.setStartBefore(audio);
    range.insertNode(newNode);
    range.deleteContents();
  }

  test(() => {
    window.onpageshow = handler;
    document.onwebkitfullscreenerror = handler;
    audio.onloadstart = handler;
    assert_true(true, 'This test must run without crash.');
  }, 'range.insertNode() should not call onpageshow event synchronously (crbug.com/660269).' );
</script>
